requirements.txt
django==3.1.1
django-mptt==0.11.0
sorl-thumbnail==12.7.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pydantic==1.8.1
gunicorn==20.1.0
